wip (fix): error at tasks page
