+++ create.html
@@ -19,7 +19,7 @@
   <script
     src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/js/bootstrap-datepicker.min.js"></script>
   <script src="./js/create.js"></script>
-  <script type="module" src="/js/TaskManager.js"></script>
+  <script src="/js/TaskManager.js" defer></script>
   <script src="./js/displayDate.js" defer></script>
 
   <title>Create | Task Planner</title>
