+++ js/TaskManager.js
@@ -1,4 +1,43 @@
-import { render } from "./tasks2";
+const htmlForTask = (task) => `
+					 <div class="col col-lg-8 col-xl-6">
+						<div class="card rounded-3 my-2">
+							<div class="card-body p-4">
+								<!-- TASK NAME -->
+								<p class="mb-2">
+									<span class="h2 me-2">${task.name}</span>
+									<span class="badge bg-info g1-badge">${task.status}</span>
+								</p>
+								<!-- ASSIGNEE -->
+								<p class="mb-2">
+									<span class="d-inline-block" tabindex="0" data-bs-toggle="tooltip" data-bs-placement="bottom" title="${task.assignedTo}">
+										<img src="https://ui-avatars.com/api/?name=${task.assignedTo}&background=random" alt="${task.assignedTo}" class="rounded-circle me-n2" width="45">
+									</span>
+								</p>
+								<p class="text-muted pb-2">27/10/2022</p>
+
+								<ul class="list-group rounded-0">
+									<!-- DUE DATE -->
+									<li class="list-group-item border-0 d-flex align-items-center ps-0">
+										<p class="mb-2">
+											<span class="h5 me-2"> Due Date </span>
+											<span class="h6">${task.dueDate}</span>
+										</p>
+									</li>
+									<!-- DESCRIPTION -->
+									<li class="list-group-item border-0 d-flex align-items-center ps-0">
+										<span class="h5 me-2"> Description </span>
+									</li>
+								</ul>
+								<!-- DESCRIPTION TEXT -->
+								<div class="form-textarea">
+									<textarea class="form-control" placeholder="Task summary" id="floatingTextarea">${task.description}
+                </textarea>
+								</div>
+							</div>
+						</div>
+						<!-- CARD END -->
+					</div>
+        `;
 
 // TASK 7
 /**
@@ -29,13 +68,21 @@ class TaskManager {
 		this.updateLocalStorage();
 	}
 
-	get tasks() {
-		return this.tasks;
-	}
+	// get tasks() {
+	// 	return this.tasks;
+	// }
 
 	removeTask(index) {
 		this.tasks.splice(index, 1);
 		this.updateLocalStorage();
+	}
+
+	renderTasks(tasks) {
+		const container = document.getElementById('tasks_container');
+
+		for (let task of tasks) {
+			container.innerHTML += htmlForTask(task);
+		}
 	}
 
 	updateLocalStorage() {
@@ -67,7 +114,13 @@ form.onsubmit = (e) => {
 
 	taskManager.addTask(task);
 
-	render(taskManager.tasks);
+	const allTasks = taskManager.tasks
+
+	console.log(allTasks);
+
+	// const getAllTasks = taskManager.tasks;
+
+	// taskManager.renderTasks(getAllTasks);
 };
 
-export default taskManager;
+// export default taskManager;
