+++ tasks.html
@@ -56,7 +56,7 @@
 
 	<!-- MAIN CONTENT -->
 	<main class="mt-5">
-		<!-- <div class="container-fluid main--container">  -->
+		<div class="container-fluid main--container"> 
 		<!-- CARD CONTAINER -->
 		<div class="container py-5 h-100">
 			<!-- ROW 1 -->
@@ -71,7 +71,8 @@
 	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
 		integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
 		crossorigin="anonymous"></script>
-	<script type="module" src="./js/TaskManager.js"></script>
+	<script src="./js/TaskManager.js"></script>
+	<script src="./js/tasks.js"></script>
 	<script src="./js/displayDate.js"></script>
 </body>
